requirements.txt
openai>=1.0.0  # Для OpenRouter API (аналогично OpenAI)
hyperliquid-python-sdk>=0.1.0  # SDK для HyperLiquid (установите через pip, если доступен; иначе используйте requests для REST API)
requests>=2.25.0
python-dotenv>=0.19.0  # Для загрузки переменных окружения
pandas>=1.3.0
numpy>=1.21.0
ta-lib>=0.4.0  # Для технического анализа
eth-account
web3
